[core_ibex,dv] Allow running RTL simulations in parallel
