+++ dv/uvm/core_ibex/sim.py
@@ -17,10 +17,12 @@
 """Regression script for running the Spike UVM testbench"""
 
 import argparse
+import logging
 import os
-import random
 import subprocess
 import sys
+
+from sim_cmd import get_simulator_cmd
 
 _CORE_IBEX = os.path.normpath(os.path.join(os.path.dirname(__file__)))
 _IBEX_ROOT = os.path.normpath(os.path.join(_CORE_IBEX, '../../..'))
@@ -35,102 +37,11 @@ try:
                  os.path.join(_RISCV_DV_ROOT, 'scripts')] +
                 sys.path)
 
-    from lib import (process_regression_list,
-                     read_yaml, run_cmd, run_parallel_cmd,
-                     setup_logging, RET_SUCCESS, RET_FAIL)
-    import logging
+    from lib import run_cmd, setup_logging, RET_SUCCESS, RET_FAIL
 
-    from ibex_config import parse_config
 
 finally:
     sys.path = _OLD_SYS_PATH
-
-
-class SeedGen:
-    '''A generator for seed values'''
-    def __init__(self, start_seed):
-        if start_seed is not None:
-            self.start_seed = start_seed
-        else:
-            # No seed specified. Pick one at random and print it out so the
-            # user can see what's going on.
-            self.start_seed = random.getrandbits(32)
-            logging.info("Random start seed chosen: {}"
-                         .format(self.start_seed))
-
-    def gen(self, iteration):
-        return self.start_seed + iteration
-
-
-def subst_opt(string, name, enable, replacement):
-    '''Substitute the <name> option in string
-
-    If enable is False, <name> is replaced by '' in string. If it is True,
-    <name> is replaced by replacement, which should be a string or None. If
-    replacement is None and <name> occurs in string, we throw an error.
-
-    '''
-    needle = '<{}>'.format(name)
-    if not enable:
-        return string.replace(needle, '')
-
-    if replacement is None:
-        if needle in string:
-            raise RuntimeError('No replacement defined for {} '
-                               '(used in string: {!r}).'
-                               .format(needle, string))
-        return string
-
-    return string.replace(needle, replacement)
-
-
-def subst_env_vars(string, env_vars):
-    '''Substitute environment variables in string
-
-    env_vars should be a string with a comma-separated list of environment
-    variables to substitute. For each environment variable, V, in the list, any
-    occurrence of <V> in string will be replaced by the value of the
-    environment variable with that name. If <V> occurs in the string but $V is
-    not set in the environment, an error is raised.
-
-    '''
-    env_vars = env_vars.strip()
-    if not env_vars:
-        return string
-
-    for env_var in env_vars.split(','):
-        env_var = env_var.strip()
-        needle = '<{}>'.format(env_var)
-        if needle in string:
-            value = os.environ.get(env_var)
-            if value is None:
-                raise RuntimeError('Cannot substitute {} in command because '
-                                   'the environment variable ${} is not set.'
-                                   .format(needle, env_var))
-            string = string.replace(needle, value)
-
-    return string
-
-
-def subst_cmd(cmd, enable_dict, opts_dict, env_vars):
-    '''Substitute options and environment variables in cmd
-
-    enable_dict should be a dict mapping names to bools. For each key, N, in
-    enable_dict, if enable_dict[N] is False, then all occurrences of <N> in cmd
-    will be replaced with ''. If enable_dict[N] is True, all occurrences of <N>
-    in cmd will be replaced with opts_dict[N].
-
-    If N is not a key in opts_dict, this is no problem unless cmd contains
-    <N>, in which case we throw a RuntimeError.
-
-    Finally, the environment variables are substituted as described in
-    subst_env_vars and any newlines are stripped out.
-
-    '''
-    for name, enable in enable_dict.items():
-        cmd = subst_opt(cmd, name, enable, opts_dict.get(name))
-
-    return subst_env_vars(cmd, env_vars).replace('\n', ' ')
 
 
 def subst_vars(string, var_dict):
@@ -140,105 +51,6 @@ def subst_vars(string, var_dict):
     for key, value in var_dict.items():
         string = string.replace('<{}>'.format(key), value)
     return string
-
-
-def get_yaml_for_simulator(simulator, yaml_path):
-    '''Read yaml at yaml_path and find entry for simulator'''
-    logging.info("Processing simulator setup file : %s" % yaml_path)
-    for entry in read_yaml(yaml_path):
-        if entry.get('tool') == simulator:
-            return entry
-
-    raise RuntimeError("Cannot find RTL simulator {}".format(simulator))
-
-
-def get_simulator_cmd(simulator, yaml_path, enables):
-    '''Get compile and run commands for the testbench
-
-    simulator is the name of the simulator to use. yaml_path is the path to a
-    yaml file describing various command line options. enables is a dictionary
-    keyed by option names with boolean values: true if the option is enabled.
-
-    Returns (compile_cmds, sim_cmd), which are the simulator commands to
-    compile and run the testbench, respectively. compile_cmd is a list of
-    strings (multiple commands); sim_cmd is a single string.
-
-    '''
-    entry = get_yaml_for_simulator(simulator, yaml_path)
-    env_vars = entry.get('env_var', '')
-
-    return ([subst_cmd(arg, enables, entry['compile'], env_vars)
-             for arg in entry['compile']['cmd']],
-            subst_cmd(entry['sim']['cmd'], enables, entry['sim'], env_vars))
-
-
-def filter_tests_by_config(cfg, test_list):
-    '''Filter out any unsupported tests from being executed.
-
-    This function will parse the set of RTL parameters required by a given
-    test (if any) and ensure that those parameters are supported by the
-    selected core config.
-
-    Doing this allows the run flow to be smarter about running regressions
-    with different configs (useful for CI flows).
-
-    Arguments:
-
-        cfg: string name of the ibex config being tested, should match a
-             config name from ibex_configs.yaml.
-
-        test_list: list of test entry objects parsed from the YAML testlist
-
-    Returns:
-
-        filtered_test_list: a list of test entry objects, filtered such that
-                            all tests incompatible with the specified ibex
-                            config have been removed.
-
-                            e.g. if the "small" config has been specified, this
-                            function will filter out all tests that require
-                            B-extension and PMP parameters
-    '''
-    filtered_test_list = []
-    config = parse_config(cfg, os.path.join(_IBEX_ROOT, "ibex_configs.yaml"))
-
-    for test in test_list:
-        if "rtl_params" in test:
-            param_dict = test['rtl_params']
-            matching_params = False
-            for p, p_val in param_dict.items():
-                config_val = config.get(p, None)
-                # Throw an error if required RTL parameters in the testlist
-                # have been formatted incorrectly (typos, wrong parameters, etc)
-                if config_val is None:
-                    logging.error('parameter % not found in config %' % (p, cfg))
-                    sys.exit(RET_FAIL)
-                else:
-                    # Ibex has some enum parameters, so as a result some tests
-                    # are able to run with several of these parameter values
-                    # (like bitmanipulation tests).
-                    # If this is the case, the testlist will specify all legal
-                    # enum values, check if any of them match the config.
-                    if isinstance(p_val, list):
-                        matching_params = (config_val in p_val)
-                    else:
-                        matching_params = (p_val == config_val)
-
-                    # If there is any parameter mismatch, we can terminate
-                    # immediately and exclude the test from being executed
-                    if not matching_params:
-                        break
-
-            if matching_params:
-                filtered_test_list.append(test)
-            else:
-                logging.info("""Cannot run test %s with config %s due to
-                                mismatching RTL parameter requirements %s""" %
-                             (test['test'], cfg, param_dict))
-        else:
-            filtered_test_list.append(test)
-
-    return filtered_test_list
 
 
 def rtl_compile(compile_cmds, output_dir, lsf_cmd, opts):
@@ -276,130 +88,6 @@ def rtl_compile(compile_cmds, output_dir, lsf_cmd, opts):
         # Note that we don't use run_parallel_cmd here: the commands in
         # compile_cmds need to be run serially.
         run_cmd(cmd)
-
-
-def get_test_sim_cmd(base_cmd, test, idx, seed, output_dir, bin_dir, lsf_cmd):
-    '''Generate the command that runs a test iteration in the simulator
-
-    base_cmd is the command to use before any test-specific substitutions. test
-    is a dictionary describing the test (originally read from the testlist YAML
-    file). idx is the test iteration (an integer) and seed is the corresponding
-    seed.
-
-    output_dir is the directory below which the test results will be written.
-    bin_dir is the directory containing compiled binaries. lsf_cmd (if not
-    None) is a string that runs bsub to submit the task on LSF.
-
-    Returns (desc, cmd, dirname) where desc is a description of the command,
-    cmd is the command to run and dirname is the directory in which to run it.
-
-    '''
-    it_cmd = subst_vars(base_cmd, {'seed': str(seed)})
-    sim_cmd = (it_cmd + ' ' + test['sim_opts'].replace('\n', ' ')
-               if "sim_opts" in test
-               else it_cmd)
-
-    test_name = test['test']
-
-    sim_dir = os.path.join(output_dir, '{}.{}'.format(test_name, seed))
-    binary = os.path.join(bin_dir, '{}_{}.bin'.format(test_name, idx))
-    desc = '{} with {}'.format(test['rtl_test'], binary)
-
-    # Do final interpolation into the test command for variables that depend on
-    # the test name or iteration number.
-    sim_cmd = subst_vars(sim_cmd,
-                         {
-                             'sim_dir': sim_dir,
-                             'rtl_test': test['rtl_test'],
-                             'binary': binary,
-                             'test_name': test_name,
-                             'iteration': str(idx)
-                         })
-
-    if not os.path.exists(binary):
-        raise RuntimeError('When computing simulation command for running '
-                           'iteration {} of test {}, cannot find the '
-                           'expected binary at {!r}.'
-                           .format(idx, test_name, binary))
-
-    if lsf_cmd is not None:
-        sim_cmd = lsf_cmd + ' ' + sim_cmd
-
-    return (desc, sim_cmd, sim_dir)
-
-
-def run_sim_commands(command_list, use_lsf):
-    '''Run the given list of commands
-
-    command_list should be a list of tuples (desc, cmd, dirname) where desc is
-    a human-readable description of the test, cmd is a command to run and
-    dirname is the directory in which to run it (which will be created if
-    necessary).
-
-    If use_lsf is true, the commands in command_list begin with something like
-    'bsub -Is'. It seems that we always use interactive bsub, so we'll have a
-    local process per job, which we track with run_parallel_cmd.
-
-    '''
-    # If we're in LSF mode, we submit all the commands 'at once', which means
-    # we have to create the output directories in advance.
-    if use_lsf:
-        cmds = []
-        for desc, cmd, dirname in command_list:
-            os.makedirs(dirname, exist_ok=True)
-            cmds.append(cmd)
-        run_parallel_cmd(cmds, timeout_s=600, exit_on_error=0,
-                         check_return_code=True)
-        return
-
-    # We're not in LSF mode, so we'll create the output directories as we go.
-    # That should make it a bit easier to see how far we got if there was an
-    # error.
-    for desc, cmd, dirname in command_list:
-        os.makedirs(dirname, exist_ok=True)
-        logging.info("Running " + desc)
-        run_cmd(cmd, timeout_s=600, exit_on_error=0,
-                check_return_code=True)
-
-
-def rtl_sim(sim_cmd, test_list,
-            start_seed, opts, output_dir, bin_dir, lsf_cmd):
-    """Run the testbench in the simulator
-
-    sim_cmd is the base command (as returned by get_simulator_cmd). This will
-    still have placeholders for test-specific arguments. test_list is a list of
-    test objects read from the testlist YAML file which gives the tests to run.
-
-    start_seed is the seed to use for the first test. This seed is incremented
-    for each subsequent test and is used to control things like random delays
-    on the bus. opts is a string of plusargs to give to the simulator.
-
-    output_dir is the output directory for simulation files (and the directory
-    in which the simulator gets run). bin_dir is the directory containing
-    binaries to be run.
-
-    If lsf_cmd is not None, it should be prefixed on each command, which will
-    be run in parallel.
-
-    """
-    logging.info("Running RTL simulation...")
-
-    sim_cmd = subst_vars(sim_cmd,
-                         {
-                             'out': output_dir,
-                             'sim_opts': opts,
-                             'cwd': _CORE_IBEX,
-                         })
-
-    # Compute a list of pairs (cmd, dirname) where cmd is the command to run
-    # and dirname is the directory in which the command should be run.
-    cmd_list = []
-    for test in test_list:
-        for i in range(test['iterations']):
-            cmd_list.append(get_test_sim_cmd(sim_cmd, test, i, start_seed + i,
-                                             output_dir, bin_dir, lsf_cmd))
-
-    run_sim_commands(cmd_list, lsf_cmd is not None)
 
 
 #TODO(udinator) - support DSim, and Riviera
@@ -442,20 +130,6 @@ def gen_cov(base_dir, simulator, lsf_cmd):
         sys.exit(RET_FAIL)
 
 
-def read_seed(arg):
-    '''Read an argument to the --start_seed command line value'''
-    try:
-        seed = int(arg)
-        if seed < 0:
-            raise ValueError('bad seed')
-        return seed
-
-    except ValueError:
-        raise argparse.ArgumentTypeError('Bad argument for --seed ({}): '
-                                         'must be a non-negative integer.'
-                                         .format(arg))
-
-
 def main():
     '''Entry point when run as a script'''
 
@@ -464,12 +138,6 @@ def main():
 
     parser.add_argument("--o", type=str, default="out",
                         help="Output directory name")
-    parser.add_argument("--testlist", help="Regression testlist",
-                        default=os.path.join(_CORE_IBEX,
-                                             'riscv_dv_extension',
-                                             'testlist.yaml'))
-    parser.add_argument("--test", type=str, default="all",
-                        help="Test name, 'all' means all tests in the list")
     parser.add_argument("--simulator", type=str, default="vcs",
                         help="RTL simulator to use (default: vcs)")
     parser.add_argument("--simulator_yaml",
@@ -477,43 +145,19 @@ def main():
                         default=os.path.join(_CORE_IBEX,
                                              'yaml',
                                              'rtl_simulation.yaml'))
-    parser.add_argument("--iss", default="spike",
-                        choices=['spike', 'ovpsim'],
-                        help="Instruction set simulator")
     parser.add_argument("-v", "--verbose", dest="verbose", action="store_true",
                         help="Verbose logging")
     parser.add_argument("--cmp_opts", type=str, default="",
                         help="Compile options for the generator")
-    parser.add_argument("--sim_opts", type=str, default="",
-                        help="Simulation options for the generator")
     parser.add_argument("--en_cov", action='store_true',
                         help="Enable coverage dump")
     parser.add_argument("--en_wave", action='store_true',
                         help="Enable waveform dump")
     parser.add_argument("--steps", type=str, default="all",
-                        help="Run steps: compile,sim,cov")
+                        help="Run steps: compile,cov")
     parser.add_argument("--lsf_cmd", type=str,
                         help=("LSF command. Run locally if lsf "
                               "command is not specified"))
-    parser.add_argument("--ibex_config", type=str,
-                        help=("Name of the Ibex config being tested. "
-                              "Should be a valid config name from "
-                              "ibex_configs.yaml"))
-
-    sg = (parser.
-          add_argument_group('Seeds and iterations',
-                             'If none of the arguments in this section are '
-                             'used, a random starting seed will be picked and '
-                             'passed as --start_seed. The number of '
-                             'iterations for each test will be read from the '
-                             'configuration.'))
-
-    sg.add_argument("--start_seed", type=read_seed, metavar='S',
-                    help=("Randomization seed for the first iteration of each "
-                          "test. Following iterations will use seeds S+1, "
-                          "S+2, and so on."))
-    sg.add_argument("--iterations", type=int,
-                    help="Override the iteration count in the test list")
 
     args = parser.parse_args()
     setup_logging(args.verbose)
@@ -527,20 +171,15 @@ def main():
 
     # Create the output directory
     output_dir = ("%s/rtl_sim" % args.o)
-    bin_dir = ("%s/instr_gen/asm_test" % args.o)
     subprocess.run(["mkdir", "-p", output_dir])
 
     steps = {
         'compile': args.steps == "all" or 'compile' in args.steps,
-        'sim': args.steps == "all" or 'sim' in args.steps,
         'cov': args.steps == "all" or 'cov' in args.steps
     }
 
-    compile_cmds = []
-    sim_cmd = ""
-    matched_list = []
-
-    if steps['compile'] or steps['sim']:
+    # Compile TB
+    if steps['compile']:
         enables = {
             'cov_opts': args.en_cov,
             'wave_opts': args.en_wave
@@ -548,22 +187,7 @@ def main():
         compile_cmds, sim_cmd = get_simulator_cmd(args.simulator,
                                                   args.simulator_yaml, enables)
 
-    # Compile TB
-    if steps['compile']:
         rtl_compile(compile_cmds, output_dir, args.lsf_cmd, args.cmp_opts)
-
-    # Run RTL simulation
-    if steps['sim']:
-        process_regression_list(args.testlist, args.test, args.iterations or 0,
-                                matched_list, _RISCV_DV_ROOT)
-        if not matched_list:
-            raise RuntimeError("Cannot find %s in %s" %
-                               (args.test, args.testlist))
-
-        matched_list = filter_tests_by_config(args.ibex_config, matched_list)
-
-        rtl_sim(sim_cmd, matched_list, args.start_seed,
-                args.sim_opts, output_dir, bin_dir, args.lsf_cmd)
 
     # Generate merged coverage directory and load it into appropriate GUI
     if steps['cov']:
